feat: Add 'operator' field to exchanges table migration
Add a new 'operator' field to the exchanges table migration with a default value of '*'. This field will be used for calculations in the WidgetCalculator component.

refactor: Update Select component in Form

Update the Select component in the Form directory to include an 'infoIcon' prop and an 'iconLabel' prop. This change allows for displaying an information icon next to the label and providing additional information to the user.

refactor: Update WidgetCalculator component

Update the WidgetCalculator component to handle the case when there is only one global exchange. Also, remove the disabled attribute from the select element and the input elements.

feat: Add 'operator' field to ExchangeConfigRates component

Add a new 'operator' field to the ExchangeConfigRates component with a default value of '*'. This field will be used for calculations in the recalculateToReceive and recalculateToSend functions.

refactor: Update Calculator utility functions

Update the Calculator utility functions to include a new 'calculateByOperator' function. This function performs calculations based on the provided operator and operands.